+++ src/db/migrations/migrations.js
@@ -11,6 +11,7 @@ import m0006 from './0006_item_price.sql';
 import m0007 from './0007_shop_total.sql';
 import m0008 from './0008_wishes.sql';
 import m0009 from './0009_pantry.sql';
+import m0010 from './0010_list_pantry.sql';
 
   export default {
     journal,
@@ -24,7 +25,8 @@ m0005,
 m0006,
 m0007,
 m0008,
-m0009
+m0009,
+m0010
     }
   }